Fix wrong render skipping logic
activeAgentId means the last agent that executed an action, not the id of the PrimaryAgent, especially not necessarily the human. While hard-coding this logic is not ideal, it is a good enough current solution.

Originally, the skip also ended the loop early and skipped the adding metadata part, causing e.g. objects to be null.